Filter IRL out of entry time distribution graph
After having put in 7h for a hackathon, it skewed all the data and made everything else nearly unreadable. In any case, I dont think IRL is useful to track here. Its nice to know how much in the aggregate I spend (achieved by the total time logged graph), but this graph is more examining how i distribute time over tasks, and working on improving that (for example, by focusing for an hour on something, rather than for varying amounts of time)