Returns last 5 objects for an area upon insert
This makes it so that, for example, you don't need to re-enter the name of the author and book again and again if you're adding a RDG task. Instead, add the book once, then just type the numcode when its returned on the next insert. Number of tasks returned is by default 5, but can be modified in the config yaml