improve prompt input for yes
At the end of the user flow to complete an entry, faereld will ask:

Is this correct? (y/n) ::

This will require the user to press "y" and "y" only to accept the
input. This pull-request changes the display (and behaviour) to:

Is this correct? (Y/n) ::

I.e. uppercase "Y" is displayed. This is a common convention in many
command line tools to present user input like this to indicate what the
default value will be. In this case, if user simply presses enter the
data will be accepted and added to the database. This will save a
keystroke everytime a user inputs data to faereld. The character "Y" and
"y" are also acceptable inputs to accept the data.