allow for hour.minute entries only with gregorian dates (#6)
I often find myself just wanting to enter a time with today's date. For
example, If I forgot to start Færeld at the beginning of the task. In
that case, I usually want to do the entry before leaving the computer
for the day. In such cases I would just like to enter two time
specifications.

This only implements this feature for Gregorian calendar. If the
implementation is deemed appropriate I will port it to Wendings.